requirements.txt
argparse==1.4.0
gitpython==2.1.14
